Fix AST representation of array accesses on fields
Given `$o->f[1]`, our parser was parsing it as an object member access
of which the member is an array access `f[1]`. However, the way that PHP
parses it, is that it is an array access on an array `$o->f`.

This commit changes our parser grammar to make it more coherent with the
PHP parses the code.