Adapt ITS expected results
The change in how we represent array accesses on fields in the AST fixes
some false positive in `S2037` SelfKeywordUsageCheck. These false
positives did appeach because field accesses of the form `self::$f[1]`
were not related correctly to their declaration as private fields (on
which the rule should not raise an issue).