SONARPHP-1118: Array accesses on fields should be represented correctly in AST (#725)
* Fix AST representation of array accesses on fields

Given `$o->f[1]`, our parser was parsing it as an object member access
of which the member is an array access `f[1]`. However, the way that PHP
parses it, is that it is an array access on an array `$o->f`.

This commit changes our parser grammar to make it more coherent with the
PHP parses the code.

* Adapt ITS expected results

The change in how we represent array accesses on fields in the AST fixes
some false positive in `S2037` SelfKeywordUsageCheck. These false
positives did appeach because field accesses of the form `self::$f[1]`
were not related correctly to their declaration as private fields (on
which the rule should not raise an issue).

* Remove unused methods

* Simplify condition in TreeFactory

* Remove unused key in PHPLexicalGrammar